등록 버튼 테스트 부분을 App테스트에서 DataInput테스트로 옮겨라
